dashboard/requirements.txt
streamlit>=1.35.0
plotly>=5.22.0
pandas>=3.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
requests>=2.32.0